Upgrade Redux, unify store creation
* Upgrade Redux
* Use the new Redux API signature for store creation
* Unify store creation for prod and dev
