+++ package.json
@@ -92,10 +92,10 @@
     "react-dnd-html5-backend": "2.0.2",
     "react-dom": "0.14.6",
     "react-dom-stream": "0.4.1",
-    "react-redux": "4.0.6",
+    "react-redux": "^4.4.0",
     "react-router": "2.0.0-rc5",
     "react-tap-event-plugin": "0.2.1",
-    "redux": "3.0.5",
+    "redux": "^3.3.1",
     "redux-form": "4.1.3",
     "redux-immutablejs": "0.0.8",
     "redux-logger": "2.3.2",
