--- src/client/storeCreator.js
@@ -1,6 +1,0 @@
-import {createStore} from 'redux';
-
-export default () =>
-  window.devToolsExtension
-    ? window.devToolsExtension()(createStore)
-    : createStore;
